crawler/requirements/requirements.txt
Flask==2.1.2
watchdog==2.1.9
hydra-core==1.1.1
omegaconf==2.1.1
pandas==1.4.2
tushare==1.2.85
protoc-gen-validate==0.4.2
SQLAlchemy==1.4.37
PyMySQL==1.0.2
protobuf==3.18.1
google-api-python-client==2.52.0
python-consul==1.1.0
netifaces==0.11.0

grpcio-tools==1.47.0
grpcio==1.47.0
grpcio-health-checking==1.47.0
grpcio-reflection==1.47.0
grpc-interceptor==0.14.2